pyproject.toml
[project]
name = "style-transfer"
version = "0.1.0"
description = ""
authors = [
    {name = "sofiadorogova",email = "113040802+sofiadorogova@users.noreply.github.com"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "openslide-python (>=1.4.1,<2.0.0)",
    "openslide-bin (>=4.0.0.6,<5.0.0.0)"
]
package-mode = false


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
